refactor into separate hooks
