+++ hooks/usePaprSalesData/index.ts
@@ -1,0 +1,1 @@
+export { usePaprSalesData } from './usePaprSalesData';
