+++ pages/tokens/[token]/net/index.tsx
@@ -6,17 +6,12 @@ import { ControllerContextProvider } from 'hooks/useController';
 import { MarketPriceProvider } from 'hooks/useLatestMarketPrice';
 import { OracleInfoProvider } from 'hooks/useOracleInfo/useOracleInfo';
 import { configProxy, SupportedToken } from 'lib/config';
-import {
-  fetchSubgraphData,
-  SubgraphController,
-  SubgraphPool,
-} from 'lib/PaprController';
+import { fetchSubgraphData, SubgraphController } from 'lib/PaprController';
 import { GetServerSideProps } from 'next';
 import { useMemo } from 'react';
 
 type ServerSideProps = {
   subgraphController: SubgraphController;
-  subgraphPool: SubgraphPool;
 };
 
 export const getServerSideProps: GetServerSideProps<ServerSideProps> = async (
@@ -50,7 +45,6 @@ export const getServerSideProps: GetServerSideProps<ServerSideProps> = async (
     props: {
       controllerAddress: config.controllerAddress,
       subgraphController: paprController,
-      subgraphPool: pool,
     },
   };
 };
